I fully overhauled the aimbot function. +Bug patch
Aimbot function now aim at the enemy that is closest from where you are currently aiming at. As you move the crosshair, the target will shift to the correct one.
There seem to be a bug that dereference null object because of the value of BoneMatrix.  When BoneMatrix is 0, the undefined behavior is gonna occur.